Change control with save reset and exit button Sucessfully added
now we have to save changed sensitivity

+++ Bot Killer/Assets/Scripts/ScensitivityCountText.cs
@@ -10,7 +10,7 @@ public class ScensitivityCountText : MonoBehaviour
    {
        if(sensitivityCountText != null)
        {
-          sensitivityCountText.SetText(optionMenu.sensitivityPercentage + "%");
+          sensitivityCountText.SetText((int)(optionMenu.sensitivityPercentage) + "%");
        }
    }
 }
